# Scheduler: reduce iteration and conversion to improve perfermance