rewrite two funcs:
1. GetNamespacesFromPodAffinityTerm func
2. PodMatchesTermsNamespaceAndSelector func

<!--  Thanks for sending a pull request!  Here are some tips for you:

1. If this is your first time, please read our contributor guidelines: https://git.k8s.io/community/contributors/guide#your-first-contribution and developer guide https://git.k8s.io/community/contributors/devel/development.md#development-guide
2. Please label this pull request according to what type of issue you are addressing, especially if this is a release targeted pull request. For reference on required PR/issue labels, read here:
https://git.k8s.io/community/contributors/devel/sig-release/release.md#issuepr-kind-label
3. Ensure you have added or ran the appropriate tests for your PR: https://git.k8s.io/community/contributors/devel/sig-testing/testing.md
4. If you want *faster* PR reviews, read how: https://git.k8s.io/community/contributors/guide/pull-requests.md#best-practices-for-faster-reviews
5. Follow the instructions for writing a release note: https://git.k8s.io/community/contributors/guide/release-notes.md
6. If the PR is unfinished, see how to mark it: https://git.k8s.io/community/contributors/guide/pull-requests.md#marking-unfinished-pull-requests
-->

**What type of PR is this?**
/kind design
/sig scheduling

**What this PR does / why we need it**:
This PR tries to reduce iteration and remove affinity term namespaces  convert to map and labelSelector convert to selector, just use term namespaces and labelSelector to match pod namespace and labels. Which can improve the scheduler perfermance:

- Required PodAntiAffinity (1.7X perfermance improvement, as below)
  - Before
  ```
  BenchmarkSchedulingPodAntiAffinity/1000Nodes/1000Pods-2         	   10000	  47852735 ns/op
  BenchmarkSchedulingPodAntiAffinity/1000Nodes/1000Pods-2         	   10000	  47919684 ns/op
  BenchmarkSchedulingPodAntiAffinity/1000Nodes/1000Pods-2         	   10000	  47430186 ns/op
  ```
  - After
  ```
  BenchmarkSchedulingPodAntiAffinity/1000Nodes/1000Pods-2         	   10000	  26860734 ns/op
  BenchmarkSchedulingPodAntiAffinity/1000Nodes/1000Pods-2         	   10000	  27151159 ns/op
  BenchmarkSchedulingPodAntiAffinity/1000Nodes/1000Pods-2         	   10000	  27231162 ns/op
  ```

**Which issue(s) this PR fixes**:
<!--
*Automatically closes linked issue when PR is merged.
Usage: `Fixes #<issue number>`, or `Fixes (paste link of issue)`.
_If PR is about `failing-tests or flakes`, please post the related issues/tests in a comment and do not use `Fixes`_*
-->
Fixes #

**Special notes for your reviewer**:
Above test result is run at a vm with 2 core cpus and 6GB memory.

**Does this PR introduce a user-facing change?**:
1.7X perfermance improvement on PodAntiAffinity

<!--
If no, just write "NONE" in the release-note block below.
If yes, a release note is required:
Enter your extended release note in the block below. If the PR requires additional action from users switching to the new release, include the string "action required".
-->
```release-note

```