@shiyan2016 2019-07-04
> It seems this PR doesn't really improve perf:
> 
> ## Without this PR
> ```
> BenchmarkSchedulingPodAffinity/5000Nodes/1000Pods-8     1000      98,540,988 ns/op
> BenchmarkSchedulingPodAffinity/5000Nodes/1000Pods-8     1000     102,982,389 ns/op
> 
> BenchmarkSchedulingPodAntiAffinity/5000Nodes/1000Pods-8     1000      34,792,657 ns/op
> BenchmarkSchedulingPodAntiAffinity/5000Nodes/1000Pods-8     1000      34,314,487 ns/op
> ```
> 
> ## With this PR
> ```
> BenchmarkSchedulingPodAffinity/5000Nodes/1000Pods-8     1000     102,279,957 ns/op
> BenchmarkSchedulingPodAffinity/5000Nodes/1000Pods-8     1000     104,050,594 ns/op
> 
> BenchmarkSchedulingPodAntiAffinity/5000Nodes/1000Pods-8     1000      35,028,057 ns/op
> BenchmarkSchedulingPodAntiAffinity/5000Nodes/1000Pods-8     1000      34,455,011 ns/op
> ```



> It seems this PR doesn't really improve perf:
> 
> ## Without this PR
> ```
> BenchmarkSchedulingPodAffinity/5000Nodes/1000Pods-8     1000      98,540,988 ns/op
> BenchmarkSchedulingPodAffinity/5000Nodes/1000Pods-8     1000     102,982,389 ns/op
> 
> BenchmarkSchedulingPodAntiAffinity/5000Nodes/1000Pods-8     1000      34,792,657 ns/op
> BenchmarkSchedulingPodAntiAffinity/5000Nodes/1000Pods-8     1000      34,314,487 ns/op
> ```
> 
> ## With this PR
> ```
> BenchmarkSchedulingPodAffinity/5000Nodes/1000Pods-8     1000     102,279,957 ns/op
> BenchmarkSchedulingPodAffinity/5000Nodes/1000Pods-8     1000     104,050,594 ns/op
> 
> BenchmarkSchedulingPodAntiAffinity/5000Nodes/1000Pods-8     1000      35,028,057 ns/op
> BenchmarkSchedulingPodAntiAffinity/5000Nodes/1000Pods-8     1000      34,455,011 ns/op
> ```

When schedule 10000 pods in a 1000 node cluster, it can improve perfermance (like PR https://github.com/kubernetes/kubernetes/pull/67788, https://github.com/kubernetes/kubernetes/pull/74041 test). See the above test result:
```
  - Before
    BenchmarkSchedulingPodAntiAffinity/1000Nodes/1000Pods-4       10000	  24375155 ns/op
  - After
   BenchmarkSchedulingPodAntiAffinity/1000Nodes/1000Pods-4        10000	   9812070 ns/op
```