@shiyan2016 2019-07-04
> @shiyan2016
> The performance numbers you posted were before the rebase, right ?
> 
> Do you mind running the benchmark on latest master and provide the comparison ?
> 
> Thanks

ok