@shiyan2016 2019-07-04
> @shiyan2016
> The performance numbers you posted were before the rebase, right ?
> 
> Do you mind running the benchmark on latest master and provide the comparison ?
> 
> Thanks

After rebase, the result are as follows:
```
  - Before
    BenchmarkSchedulingPodAntiAffinity/1000Nodes/1000Pods-4         10000	  25698313 ns/op
  - After
    BenchmarkSchedulingPodAntiAffinity/1000Nodes/1000Pods-4        10000	  10649081 ns/op
```