@shiyan2016 2019-07-05
/test pull-kubernetes-integration