@wgliang 2019-07-10
/test pull-kubernetes-bazel-test