@shiyan2016 2019-07-17
@wgliang @Huang-Wei  I have done the rebase. Pls have a look.Thanks very much.