@Huang-Wei 2019-07-17
@shiyan2016 I'm having a quite limited time reviewing PRs this week. Sry for that.

After a quick look, my biggest concerns are:

1. Performance. I'm still not 100% sure we can get enough perf gain.
1. This PR changed an interface to a concrete fixed type. Tightening the arguments to a fixed type doesn't sound like a good idea, esp. in terms of testability and flexibility for possible future changes.
1. Added helper code is not equipped with enough unit tests.
1. Those helper code may be moved into apimachenary packages for reusability.

I will do a close look at perf next week and get back.