@shiyan2016 2019-07-17
/test pull-kubernetes-verify