@shiyan2016 2019-07-18
/test pull-kubernetes-e2e-gce-device-plugin-gpu